# docs/api/external-v1.openapi.yaml
openapi: 3.1.0
info:
  title: bluDM External Authoring API
  version: 1.0.0
  description: Versioned REST adapter over the same application services used by bluDM MCP.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://{host}
    description: Replace the host with the public or private bluDM deployment hostname.
    variables:
      host:
        default: bludm.invalid
security:
  - bearerAuth: []
...
paths:
...
  /api/external/v1/campaigns/{campaignId}/players:
    get:
      operationId: listPlayers
      summary: List campaign player characters.
      tags: [discovery]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }]
      responses: { "200": { $ref: "#/components/responses/PlayerPage" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/players/{playerId}:
    get:
      operationId: getPlayer
      summary: Get one campaign player character.
      tags: [discovery]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/PlayerId" }]
      responses: { "200": { $ref: "#/components/responses/Player" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/locations:
    get:
      operationId: listLocations
      summary: List and filter Campaign World locations.
      tags: [world]
      parameters:
        - { $ref: "#/components/parameters/CampaignId" }
        - { name: q, in: query, schema: { type: string }, description: Case-insensitive name and notes filter. }
        - { name: type, in: query, schema: { type: string } }
        - { name: parentLocationId, in: query, schema: { type: string }, description: Empty string selects root locations. }
        - { name: status, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses: { "200": { $ref: "#/components/responses/LocationPage" }, default: { $ref: "#/components/responses/Error" } }
    post:
      operationId: createLocation
      summary: Create a Campaign World location.
      tags: [world]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/RequiredIdempotencyKey" }]
      requestBody: { $ref: "#/components/requestBodies/LocationCommand" }
      responses: { "201": { $ref: "#/components/responses/LocationWrite" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/locations/{locationId}:
    get:
      operationId: getLocation
      summary: Get a location and its related campaign context.
      tags: [world]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/LocationId" }]
      responses: { "200": { $ref: "#/components/responses/LocationContext" }, default: { $ref: "#/components/responses/Error" } }
    patch:
      operationId: updateLocation
      summary: Update a location with optimistic concurrency.
      tags: [world]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/LocationId" }, { $ref: "#/components/parameters/RequiredIdempotencyKey" }]
      requestBody: { $ref: "#/components/requestBodies/UpdateLocationCommand" }
      responses: { "200": { $ref: "#/components/responses/LocationWrite" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/world-graph:
    get:
      operationId: getWorldGraph
      summary: Get the campaign location and connection graph.
      tags: [world]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }]
      responses: { "200": { $ref: "#/components/responses/WorldGraph" }, default: { $ref: "#/components/responses/Error" } }
...
  /api/external/v1/campaigns/{campaignId}/library/creatures/{creatureId}:
    get:
      operationId: getCreature
      summary: Get a complete creature record.
      tags: [discovery]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/CreatureId" }]
      responses: { "200": { $ref: "#/components/responses/CreatureDetails" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/library/entries:
    get:
      operationId: searchLibrary
      summary: Search spells, items, rules, and other library entries.
      tags: [discovery]
      parameters:
        - { $ref: "#/components/parameters/CampaignId" }
        - { name: type, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
      responses: { "200": { $ref: "#/components/responses/LibrarySearch" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/library/entries/{entryId}:
    get:
      operationId: getLibraryEntry
      summary: Get one typed library entry.
      tags: [discovery]
      parameters:
        - { $ref: "#/components/parameters/CampaignId" }
        - { name: entryId, in: path, required: true, schema: { type: string } }
        - { name: type, in: query, required: true, schema: { type: string } }
        - { name: source, in: query, schema: { enum: [user, standard] } }
      responses: { "200": { $ref: "#/components/responses/LibraryEntry" }, default: { $ref: "#/components/responses/Error" } }
...
  /api/external/v1/campaigns/{campaignId}/encounters/{encounterId}:
    get:
      operationId: getEncounter
      summary: Get an encounter, roster, and difficulty evidence.
      tags: [encounters]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/EncounterId" }]
      responses: { "200": { $ref: "#/components/responses/EncounterDetails" }, default: { $ref: "#/components/responses/Error" } }
    patch:
      operationId: updateEncounter
      summary: Apply a targeted encounter update.
      tags: [encounters]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/EncounterId" }, { $ref: "#/components/parameters/RequiredIdempotencyKey" }]
      requestBody: { $ref: "#/components/requestBodies/UpdateEncounter" }
      responses: { "200": { $ref: "#/components/responses/EncounterWrite" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/generation/encounter-evaluation:
    post:
      operationId: evaluateEncounter
      summary: Evaluate an explicit encounter roster without saving.
      tags: [encounters]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }]
      requestBody: { $ref: "#/components/requestBodies/EvaluateEncounter" }
      responses: { "200": { $ref: "#/components/responses/DifficultyEvidence" }, default: { $ref: "#/components/responses/Error" } }
...
  /api/external/v1/campaigns/{campaignId}/npcs:
    post:
      operationId: createNpc
      summary: Create a campaign-linked custom NPC.
      tags: [world]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/RequiredIdempotencyKey" }]
      requestBody: { $ref: "#/components/requestBodies/NPCCommand" }
      responses: { "201": { $ref: "#/components/responses/NPCWrite" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/npcs/{npcId}:
    patch:
      operationId: updateNpc
      summary: Update an NPC with optimistic concurrency.
      tags: [world]
      parameters:
        - { $ref: "#/components/parameters/CampaignId" }
        - { name: npcId, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/RequiredIdempotencyKey" }
      requestBody: { $ref: "#/components/requestBodies/UpdateNPCCommand" }
      responses: { "200": { $ref: "#/components/responses/NPCWrite" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/npc-location-links:
    post:
      operationId: linkNpcToLocation
      summary: Link an NPC to a campaign location.
      tags: [world]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/RequiredIdempotencyKey" }]
      requestBody: { $ref: "#/components/requestBodies/NPCLinkCommand" }
      responses: { "201": { $ref: "#/components/responses/NPCLinkWrite" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/location-links:
    post:
      operationId: createLocationLink
      summary: Create a relationship between two locations.
      tags: [world]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/RequiredIdempotencyKey" }]
      requestBody: { $ref: "#/components/requestBodies/LocationLinkCommand" }
      responses: { "201": { $ref: "#/components/responses/LocationLinkWrite" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/content/changes/preview:
    post:
      operationId: previewCampaignChanges
      summary: Validate and preview a connected campaign change set.
      tags: [world]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }]
      requestBody: { $ref: "#/components/requestBodies/CampaignChanges" }
      responses: { "200": { $ref: "#/components/responses/CampaignChangesPreview" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/content/changes/apply:
    post:
      operationId: applyCampaignChanges
      summary: Apply an approved campaign change set transactionally.
      tags: [world]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/RequiredIdempotencyKey" }]
      requestBody: { $ref: "#/components/requestBodies/CampaignChanges" }
      responses: { "200": { $ref: "#/components/responses/CampaignChangesApplied" }, default: { $ref: "#/components/responses/Error" } }
  /api/external/v1/campaigns/{campaignId}/roll-tables:
    get:
      operationId: listRollTables
      summary: List campaign roll tables.
      tags: [advanced]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }]
      responses: { "200": { $ref: "#/components/responses/RollTablePage" }, default: { $ref: "#/components/responses/Error" } }
    post:
      operationId: createRollTable
      summary: Create a campaign roll table.
      tags: [advanced]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/RequiredIdempotencyKey" }]
      requestBody: { $ref: "#/components/requestBodies/RollTableCommand" }
      responses: { "201": { $ref: "#/components/responses/RollTableWrite" }, default: { $ref: "#/components/responses/Error" } }
...
  /api/external/v1/campaigns/{campaignId}/journeys:
    post:
      operationId: createJourney
      summary: Save a campaign journey.
      tags: [advanced]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }, { $ref: "#/components/parameters/RequiredIdempotencyKey" }]
      requestBody: { $ref: "#/components/requestBodies/JourneyCommand" }
      responses: { "201": { $ref: "#/components/responses/JourneyWrite" }, default: { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Scoped bludm_v1_ API token or configured OIDC JWT access token.
  parameters:
    CampaignId: { name: campaignId, in: path, required: true, schema: { type: string, format: uuid } }
    PlayerId: { name: playerId, in: path, required: true, schema: { type: string, format: uuid } }
    LocationId: { name: locationId, in: path, required: true, schema: { type: string, format: uuid } }
    EncounterId: { name: encounterId, in: path, required: true, schema: { type: string, format: uuid } }
    CreatureId: { name: creatureId, in: path, required: true, schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
    Cursor: { name: cursor, in: query, schema: { type: string }, description: Opaque cursor from the previous page. }
    RequiredIdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 256 } }
...
  responses:
...
    PlayerPage: { description: Compact party discovery page., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/PlayerPage" } } } }
    Player: { description: Full player character record., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/Player" } } } }
    LocationPage: { description: Location discovery page., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/LocationPage" } } } }
    LocationContext: { description: Location context and relationships., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/LocationContext" } } } }
    LocationWrite: { description: Saved location with write evidence., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/LocationWriteResult" } } } }
    WorldGraph: { description: Campaign location graph., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/WorldGraph" } } } }
...
    CreatureDetails: { description: "Creature details, actions, and spellcasting.", content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/CreatureDetails" } } } }
    LibrarySearch: { description: Library search page., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/LibrarySearch" } } } }
    LibraryEntry: { description: Typed library entry., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/LibraryEntry" } } } }
...
    EncounterDetails: { description: Encounter roster and server-calculated difficulty., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/EncounterDetails" } } } }
    EncounterWrite: { description: Saved authored encounter with write evidence., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/EncounterWriteResult" } } } }
    DifficultyEvidence: { description: Server-authoritative encounter difficulty evidence., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/DifficultyEvidence" } } } }
...
    NPCWrite: { description: Saved NPC with write evidence., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/NPCWriteResult" } } } }
    NPCLinkWrite: { description: Saved NPC placement with write evidence., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/NPCLinkWriteResult" } } } }
    LocationLinkWrite: { description: Saved location link with write evidence., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/LocationLinkWriteResult" } } } }
    CampaignChangesPreview: { description: Normalized connected change-set approval preview., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/CampaignChangesPreview" } } } }
    CampaignChangesApplied: { description: Transactional connected change-set result., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/CampaignChangesApplied" } } } }
    RollTablePage: { description: Campaign roll tables., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/RollTablePage" } } } }
    RollTableWrite: { description: Saved roll table with write evidence., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/RollTableWriteResult" } } } }
...
    JourneyWrite: { description: Saved journey with write evidence., content: { application/json: { schema: { $ref: "./external-v1.schemas.yaml#/components/schemas/JourneyWriteResult" } } } }
...
    Error:
      description: Typed external API error.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            error: { code: conflict, message: encounter revision changed, requestId: req-123 }
  requestBodies:
    LocationCommand:
      required: true
      content: { application/json: { schema: { $ref: "#/components/schemas/LocationCommand" } } }
    UpdateLocationCommand:
      required: true
      content: { application/json: { schema: { $ref: "#/components/schemas/UpdateLocationCommand" } } }
...
    UpdateEncounter:
      required: true
      content: { application/json: { schema: { $ref: "#/components/schemas/UpdateEncounter" } } }
    EvaluateEncounter:
      required: true
      content: { application/json: { schema: { $ref: "#/components/schemas/EvaluateEncounter" } } }
...
    CampaignChanges:
      required: true
      content: { application/json: { schema: { $ref: "#/components/schemas/CampaignChanges" } } }
    NPCCommand:
      required: true
      content: { application/json: { schema: { $ref: "#/components/schemas/NPCCommand" } } }
    UpdateNPCCommand:
      required: true
      content: { application/json: { schema: { $ref: "#/components/schemas/UpdateNPCCommand" } } }
    NPCLinkCommand:
      required: true
      content: { application/json: { schema: { $ref: "#/components/schemas/NPCLinkCommand" } } }
    LocationLinkCommand:
      required: true
      content: { application/json: { schema: { $ref: "#/components/schemas/LocationLinkCommand" } } }
    RollTableCommand:
      required: true
      content: { application/json: { schema: { $ref: "#/components/schemas/RollTableCommand" } } }
...
    JourneyCommand:
      required: true
      content: { application/json: { schema: { $ref: "#/components/schemas/JourneyCommand" } } }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { enum: [unauthorized, forbidden, not_found, validation_failed, conflict, idempotency_conflict, rate_limited, timeout, unsupported, internal_error] }
            message: { type: string }
            details: { type: object }
            requestId: { type: string }
          additionalProperties: false
      additionalProperties: false
    LocationCommand:
      type: object
      required: [name, locationType]
      properties:
        idempotencyKey: { type: string }
        expectedUpdatedAt: { type: string, format: date-time }
        parentLocationId: { type: string, format: uuid }
        name: { type: string, minLength: 1 }
        locationType: { type: string }
        summary: { type: string }
        notes: { type: string }
        publicNotes: { type: string }
        dmNotes: { type: string }
        tags: { type: array, items: { type: string } }
        status: { type: string }
      additionalProperties: false
    UpdateLocationCommand:
      allOf:
        - { $ref: "#/components/schemas/LocationCommand" }
        - { type: object, required: [expectedUpdatedAt] }
...
    UpdateEncounter:
      type: object
      required: [idempotencyKey, expectedRevision]
      properties:
        idempotencyKey: { type: string, minLength: 1 }
        expectedRevision: { type: integer, minimum: 1 }
        name: { type: string, minLength: 1 }
        description: { type: string }
        status: { enum: [planned, completed, skipped] }
        locationId: { type: string, format: uuid, description: An empty string clears the link. }
        location: { type: string }
        roomNumber: { type: string }
        lootNotes: { type: string }
        replaceRoster: { type: boolean, default: false }
        combatants:
          type: array
          maxItems: 100
          description: Full replacement roster; only valid with replaceRoster=true.
          items: { $ref: "#/components/schemas/EncounterCombatantCommand" }
        addCombatants:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/EncounterCombatantCommand" }
        updateCombatants:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/EncounterCombatantPatch" }
        removeCombatantIds:
          type: array
          maxItems: 100
          uniqueItems: true
          items: { type: string, format: uuid }
      additionalProperties: false
      examples:
        - idempotencyKey: encounter-quantity-edit-01
          expectedRevision: 2
          updateCombatants:
            - combatantId: 33333333-3333-4333-8333-333333333333
              displayName: Dire Wolf pack
...
    EvaluateEncounter:
      type: object
      required: [allCampaignPlayers, enemies, requestedDifficulty]
      properties:
        allCampaignPlayers: { type: boolean }
        playerIds: { type: array, items: { type: string, format: uuid } }
        enemies:
          type: array
          minItems: 1
          items:
            type: object
            required: [creatureId, quantity]
            properties:
              creatureId: { type: string }
              quantity: { type: integer, minimum: 1, maximum: 50 }
            additionalProperties: false
        requestedDifficulty: { enum: [easy, medium, hard, deadly] }
      additionalProperties: false
      examples:
        - allCampaignPlayers: false
          playerIds: [11111111-1111-4111-8111-111111111111]
          enemies:
            - creatureId: standard:dire-wolf
              quantity: 3
          requestedDifficulty: medium
...
    NPCCommand:
      type: object
      required: [name, size, creatureType, armorClass, hitPoints, hitDice, challengeRating, xp, statBlock]
      properties:
        idempotencyKey: { type: string }
        expectedUpdatedAt: { type: string, format: date-time }
        name: { type: string, minLength: 1 }
        description: { type: string }
        size: { type: string }
        creatureType: { type: string }
        alignment: { type: string }
        armorClass: { type: integer, minimum: 1 }
        hitPoints: { type: integer, minimum: 1 }
        hitDice: { type: string }
        challengeRating: { type: string }
        xp: { type: integer, minimum: 0 }
        avatarUrl: { type: string }
        statBlock: { type: object }
        disposition: { type: string }
      additionalProperties: false
    UpdateNPCCommand:
      allOf:
        - { $ref: "#/components/schemas/NPCCommand" }
        - { type: object, required: [expectedUpdatedAt] }
    NPCLinkCommand:
      type: object
      required: [idempotencyKey, creatureId, locationId]
      properties:
        idempotencyKey: { type: string, minLength: 1 }
        creatureId: { type: string, format: uuid }
        locationId: { type: string, format: uuid }
        linkType: { type: string }
        visibility: { type: string }
        notes: { type: string }
      additionalProperties: false
    LocationLinkCommand:
      type: object
      required: [idempotencyKey, sourceLocationId, targetLocationId, linkType]
      properties:
        idempotencyKey: { type: string, minLength: 1 }
        sourceLocationId: { type: string, format: uuid }
        targetLocationId: { type: string, format: uuid }
        linkType: { type: string }
        label: { type: string }
        direction: { type: string }
        visibility: { type: string }
        notes: { type: string }
      additionalProperties: false
...